Revert "Update data type"
This reverts commit 2a0a2b7ac2aeb56352b40bdcfff67fb3d9471313.

+++ application.go
@@ -147,7 +147,7 @@ func HandleJSONUpload(c *gin.Context) {
 		employee := Employee{FirstName: obj["firstName"].(string),
 			LastName:     obj["lastName"].(string),
 			AnnualSalary: obj["annualSalary"].(float64),
-			PaymentMonth: int(obj["paymentMonth"].(float64)),
+			PaymentMonth: obj["paymentMonth"].(int),
 			SuperRate:    obj["superRate"].(float64)}
 		payslip := GenerateRESTPayslip(employee)
 		payslips = append(payslips, payslip)
